Pass queues into downloadWithArgs function
