+++ minis/downloader/__init__.py
@@ -1,3 +1,4 @@
 from .downloader import *
 from .show_info import *
 from .extract import *
+from .metadata import *
